requirements.txt
mkdocs-material
mkdocs-static-i18n
markdown-callouts
python-dotenv
deepl